+++ src/private.h
@@ -16,7 +16,7 @@ void                rumble_modules_load(masterHandle *master);
 void                rumble_database_load(masterHandle *master);
 masterHandle        *rumble_get_master(void);
 void                rumble_tag_file(FILE *fp, const char *host, const char *fid, const char *usr, const char *dmn);
-uint32_t            rumble_copy_mail(masterHandle *master, const char *fid, const char *usr, const char *dmn, char **pfid);
+size_t              rumble_copy_mail(masterHandle *master, const char *fid, const char *usr, const char *dmn, char **pfid);
 rumble_readerwriter *rumble_rw_init(void);
 void                rumble_rw_start_read(rumble_readerwriter *rrw);
 void                rumble_rw_stop_read(rumble_readerwriter *rrw);
